requirements.txt
Flask
twilio
openai
gunicorn
google-generativeai
